Resource Manager Rewrite
Resource Entries are added with the Scriptable Object, not an index in a list of them. I have no idea what was going through my mind when I made the first implementation.

A complete rewrite is required if I want to use this again in the future. Not too difficult, much of the code transfers over.